tree-optimization/106513 - fix mistake in bswap symbolic number shifts
This fixes a mistake in typing a local variable in the symbolic
shift routine.

	PR tree-optimization/106513
	* gimple-ssa-store-merging.c (do_shift_rotate): Use uint64_t
	for head_marker.

	* gcc.dg/torture/pr106513.c: New testcase.

(cherry picked from commit f675afa4eeac9910a2c085a95aa04d6d9f2fd8d6)

+++ gcc/gimple-ssa-store-merging.c
@@ -263,7 +263,7 @@ do_shift_rotate (enum tree_code code,
 		 int count)
 {
   int i, size = TYPE_PRECISION (n->type) / BITS_PER_UNIT;
-  unsigned head_marker;
+  uint64_t head_marker;
 
   if (count < 0
       || count >= TYPE_PRECISION (n->type)
